[SWM-206] Refactor : components & constants directory
