+++ src/constants/lectureDetail/lectureDetail.ts
